+++ Changelog.md
@@ -2,6 +2,7 @@
 
 ## Version 2.2
 
+- [B.9] Updated ``/deleteFrom`` command to the new API version
 - [B.9] Updated ``/shuffle`` command to the new API version
 - [B.8] Updated ``/get`` command to the new API version
 - [B.7] Updated ``/setMode`` command to the new API version